pyproject.toml
[project]
name = "ziyofat-day"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "fastapi[all,standart]>=0.129.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "ruff>=0.15.2",
    "sqlalchemy>=2.0.46",
    "starlette-admin>=0.16.0",
    "uvicorn>=0.41.0",
]
